+++ doc/changelog.md
@@ -27,6 +27,7 @@
 - log mint errors
 - Fixed denomination issue in CLI POKT to uPOKT
 - Updated PopModel so no empty body `{}` is needed in RPC
+- Added evidence type to proof message
 - Changed SendClaimTx to use sessionContext for supportedBlockchains
 - Refactor ABCI BeginBlock for app module, avoid looping twice on mature applications
 - Swapped ExecuteProof and SetReceipt order on handleProofMsg func